nft-generator/requirements.txt
aiohttp==3.7.3
pip-chill==1.0.0
python-socketio==5.0.4
requests==2.25.1
